+++ example/src/testConstants.ts
@@ -1,16 +1,7 @@
 const LOCAL_URL = 'http://localhost:19006/';
 
-const EXAMPLE_CONTENT = [
-  'Hello, *world*!',
-  'https://expensify.com',
-  '# header1',
-  '> blockquote',
-  '`inline code`',
-  '```\ncodeblock\n```',
-  '@here',
-  '@someone@swmansion.com',
-  '#mention-report',
-].join('\n');
+const EXAMPLE_CONTENT =
+  '!(https://images.unsplash.com/photo-1597431793715-b4b71ddb5670?crop=entropy&cs=tinysrgb&fit=crop&fm=jpg&h=800&ixid=MnwxfDB8MXxyYW5kb218MHx8fHx8fHx8MTcxNDQ3OTI3Nw&ixlib=rb-4.0.3&q=80&utm_campaign=api-credit&utm_medium=referral&utm_source=unsplash_source&w=1900)';
 
 const INPUT_ID = 'MarkdownInput_Example';
 const INPUT_HISTORY_DEBOUNCE_TIME_MS = 150;
